pyproject.toml
[project]
name = "kakehashi"
version = "0.1.0"
description = "ES-DE 日本語メタデータ管理ツール"
requires-python = ">=3.10"
dependencies = [
    "tkcalendar>=1.6.1",
    "pillow>=10.0.0",
    "paramiko>=3.0.0",
    "numpy>=1.24.0",
]
